Added cli_autogen_common.py, added separate dir for each cli_autogen test
