+++ tests/cli_autogen_input/yang_parser_test/assert_dictionaries.py
