+++ tests/cli_autogen_input/yang_parser_test/sonic-1-object-container.yang
