+++ tests/cli_autogen_input/yang_parser_test/sonic-2-lists.yang
