+++ tests/cli_autogen_input/yang_parser_test/sonic-choice-complex.yang
