+++ tests/cli_autogen_input/yang_parser_test/sonic-dynamic-object-complex-1.yang
