+++ tests/cli_autogen_input/yang_parser_test/sonic-grouping-2.yang
