+++ tests/cli_autogen_input/yang_parser_test/sonic-static-object-complex-1.yang
