+++ tests/cli_autogen_input/yang_parser_test/sonic-static-object-complex-2.yang
